Fix closures inside lambdas
The problem was that the `x` in the test was marked as a free variable:

    (labels
      ((label_1 (code () (x) (let ([x 1]) (tailcall (closure label_0 x)))))
        (label_0 (code () (x) x)))
      ()
      (funcall (closure label_1 x)))

That lead to the outer lambda requiring the `x` to be turned into a
closure, even though the `x` comes from the `let` covering the inner
`lambda`.

The problem was that, when annotating free variables, we didn't filter
out the bindings introduced by the let names. Those are reported as free
variables by the expressions in the let-body. But we must not report
them _up_ as such.